feat(memory_buffer): allow buffer spillover
Adds the 'spillover' boolean field to the
'memory' buffer component (default false). When
true, any incoming message batch that will put
the buffer above its configured 'limit' will be
dropped.

The motivation for this is to support input
sources that do not gracefully handle
backpressure and where the desired delivery
guarantees are best effort.